% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicPCA.R
\name{basicPCA}
\alias{basicPCA}
\title{Executa Analise de Componentes Principais (PCA)}
\usage{
basicPCA(data, scale = TRUE, center = TRUE)
}
\arguments{
\item{data}{Um data.frame ou tibble contendo apenas colunas numericas.}

\item{scale}{Logical. Se TRUE, padroniza as variaveis (desvio padrao = 1).}

\item{center}{Logical. Se TRUE, centraliza as variaveis na media.}

\item{remove_na}{Logical. Se TRUE, remove linhas contendo NA antes da PCA.}
}
\value{
Um objeto da classe `prcomp` contendo os resultados da PCA.
}
\description{
Esta funcao realiza uma PCA utilizando `prcomp`, com opcoes para
centralizacao e escalonamento das variaveis.
}
