% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plota o grafico de PCA (PC1 x PC2)}
\usage{
plotPCA(pca_result, grupos = NULL, rotulos = NULL)
}
\arguments{
\item{pca_result}{Objeto resultante de prcomp().}

\item{grupos}{Um vetor com fatores de agrupamento (ex.: condicoes).
Deve ter o mesmo numero de linhas que pca_result$x.}

\item{rotulos}{Vetor com nomes/samples para mostrar no grafico.}
}
\value{
Um grafico ggplot2.
}
\description{
Plota o grafico de PCA (PC1 x PC2)
}
