% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicPCA.R
\name{basicPCA}
\alias{basicPCA}
\title{Executa Análise de Componentes Principais (PCA)}
\usage{
basicPCA(dados, scale = TRUE, center = TRUE)
}
\arguments{
\item{dados}{Um data.frame ou tibble contendo apenas colunas numéricas.}

\item{scale}{Logical. Se TRUE, padroniza as variáveis (desvio padrão = 1).}

\item{center}{Logical. Se TRUE, centraliza as variáveis na média.}

\item{remove_na}{Logical. Se TRUE, remove linhas contendo NA antes da PCA.}
}
\value{
Um objeto da classe `prcomp` contendo os resultados da PCA.
}
\description{
Esta função realiza uma PCA utilizando `prcomp`, com opções para
centralização e escalonamento das variáveis.
}
\examples{
pca <- minha_pca(iris[, 1:4])
}
