% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{Plota o gráfico de PCA (PC1 x PC2)}
\usage{
plotPCA(pca_result, grupos = NULL, rotulos = NULL)
}
\arguments{
\item{pca_result}{Objeto resultante de prcomp().}

\item{grupos}{Um vetor com fatores de agrupamento (ex.: condições).
Deve ter o mesmo número de linhas que pca_result$x.}

\item{rotulos}{Vetor com nomes/samples para mostrar no gráfico.}
}
\value{
Um gráfico ggplot2.
}
\description{
Plota o gráfico de PCA (PC1 x PC2)
}
